// src/frontend-applications/agent-portal/modules/import-export/model/ImportExportTemplateRunProperty.ts
/**
 * Copyright (C) 2006-2024 KIX Service Software GmbH, https://www.kixdesk.com
 * --
 * This software comes with ABSOLUTELY NO WARRANTY. For details, see
 * the enclosed file LICENSE for license information (GPL3). If you
 * did not receive this file, see https://www.gnu.org/licenses/gpl-3.0.txt.
 * --
 */

export enum ImportExportTemplateRunProperty {

    ID = 'ID',

    TEMPLATE_ID = 'TemplateID',

    TYPE = 'Type',

    SUCCESS_COUNT = 'SuccessCount',

    FAIL_COUNT = 'FailCount',

    STATE_ID = 'StateID',

    STATE = 'State',

    START_TIME = 'StartTime',

    END_TIME = 'EndTime',

    DSFDS = 'CreateBy',

    LIST_NUMBER = 'listNumber',

    EXEC_IMPORT_CONTENT = 'ImportFileContent'

}